+++ examples/multiple-queue.js
@@ -9,8 +9,8 @@ const HCCrawler = require('headless-chrome-crawler');
       console.log(`Got ${result.result.title} for ${result.options.url}.`);
     }),
   });
-  crawler.queue('https://example.com/'); // Queue a request
-  crawler.queue(['https://example.net/', { url: 'https://example.org/' }]); // Queue multiple requests in different styles
+  await crawler.queue('https://example.com/'); // Queue a request
+  await crawler.queue(['https://example.net/', { url: 'https://example.org/' }]); // Queue multiple requests in different styles
   await crawler.onIdle();
   await crawler.close();
 })();